notification_service/requirements.txt
Flask
flask-cors
pika
python-socketio
eventlet
